refactor(drafter): extract step page renderers from routes.py
Move the 13 Step-1..7 wizard page renderers plus the
``_find_latest_job`` DB helper out of ``app/drafter/routes.py`` and
into a new ``app/drafter/_step_renderers.py``. ``routes.py`` shrinks
from 2471 to 1615 lines; the new module is 951 lines focused on pure
rendering.

Moved: ``_step_1_page``, ``_step_1_content``, ``_find_latest_job``,
``_step_2_page``, ``_step_3_page``, ``_research_category_card``,
``_step_4_page``, ``_render_clause_card``, ``_step_5_page``,
``_step_6_page``, ``_step_7_page``, ``_step_waiting_page``,
``_step_error_page``.

To keep existing test patches (``patch("app.drafter.routes.X")``)
working without modifying any test file, all 13 names are re-imported
at the top of ``routes.py`` so they remain attributes of the routes
module. A small ``_routes_attr(name)`` call-time indirection in the
renderers resolves ``_find_latest_job``, ``decrypt_text``, and
``_step_tracker`` against ``app.drafter.routes`` at call time, so a
patch on the routes module still takes effect inside the renderers.

Tests: 96 drafter tests pass; full suite 2454 pass / 25 skipped.
ruff and pyright clean. Zero test files modified.
